alem2006

+++ main.py
@@ -99,10 +99,10 @@ class runModel:
         data = data_preprocessor.get_preprocessed_data()
 
         X_train, X_test, y_train, y_test = self.split_normalize_XY(data)
-        tensor_X_train = torch.tensor(X_train, dtype=torch.float16)
-        tensor_y_train = torch.tensor(y_train, dtype=torch.float16)
-        tensor_X_test = torch.tensor(X_test, dtype=torch.float16)
-        tensor_y_test = torch.tensor(y_test, dtype=torch.float16)
+        tensor_X_train = torch.tensor(X_train, dtype=torch.float)
+        tensor_y_train = torch.tensor(y_train, dtype=torch.float)
+        tensor_X_test = torch.tensor(X_test, dtype=torch.float)
+        tensor_y_test = torch.tensor(y_test, dtype=torch.float)
 
         train_dataset = TensorDataset(tensor_X_train, tensor_y_train)
         test_dataset = TensorDataset(tensor_X_test, tensor_y_test)
@@ -110,7 +110,7 @@ class runModel:
         dataloader_train_set = DataLoader(train_dataset, batch_size=self.batch_size, shuffle=True)
         dataloader_test_set = DataLoader(test_dataset, batch_size=self.batch_size)
 
-        self.cnn_model = CNN()
+        self.cnn_model = CNN().float()
         self.CNN_loss_func = nn.MSELoss()
         self.CNN_optimizer = optim.Adam(self.cnn_model.parameters(), lr = self.learning_rate)
         train(self.cnn_model, dataloader_train_set, self.CNN_loss_func, self.CNN_optimizer, self.epochs)
